Add Comments

+++ src/Controller/ApiController.php
@@ -43,6 +43,7 @@ class ApiController extends AbstractController
         $classes=$em->getRepository('App:Classes')->findAll();
 
         $jsonContent = $serializer->serialize($classes, 'json');
+
         $response = new JsonResponse();
         $response->headers->set('Content-Type', 'application/text');
         $response->headers->set('Access-Control-Allow-Origin', '*');
@@ -53,8 +54,9 @@ class ApiController extends AbstractController
         
     }
     /**
-     * @Route("api/classes",name="add_api_classes", methods={"PUT", "OPTIONS"})
-     * @Route("api/classes/{id}/edit", name="api_editClass", methods={"POST", "OPTIONS"})
+     * @Route("api/classes",name="api_add_classes", methods={"PUT", "OPTIONS"})
+     * @Route("api/classes/{id}/edit", name="api_edit_class", methods={"POST", "OPTIONS"})
+     * @param int $id id de la classe, mode Edit
      */
     public function APIaddClass(Classes $class = null, Request $request, $id=null)
     {
@@ -67,23 +69,27 @@ class ApiController extends AbstractController
             $response->headers->set('Access-Control-Allow-Headers', 'Content-Type', true);
             return $response;   
         }
-        if ($id)
-        {
-            $entityManager=$this->getDoctrine()->getManager();
-            $em=$entityManager->getRepository(Classes::class)->find($id);
-            $nameClass=$em->getNameClass();
-        }
-        $routeEdit=true; 
+        $editMode=true; 
+        //Si aucune classe à modifier n'est fournie -> edit Mode = false
         if(!$class)
         {
             $class = new Classes();
-            $routeEdit=false; 
+            $editMode=false; 
         }
+        if ($editMode==true)
+        {
+            $entityManager=$this->getDoctrine()->getManager();
+            //Enregistre dans $em la classe de l'id fournie en edit Mode
+            $em=$entityManager->getRepository(Classes::class)->find($id);
+            //Cette variable $nameClass est nécessaire pour modifier tous les étudiants
+            //qui sont liés à cet nom de classe
+            $nameClass=$em->getNameClass();
+        }
+        //Création du formulaire
         $data = json_decode($request->getContent(),true);
         $form = $this->createForm(ClassesType::class, $class);
-
         $form->submit($data);
-    
+        
         if (false === $form->isValid()) {
             $response=new JsonResponse(
                 [
@@ -97,13 +103,15 @@ class ApiController extends AbstractController
             $response->headers->set("Access-Control-Allow-Methods", "GET, PUT, POST, DELETE, OPTIONS");
             $response->headers->set('Access-Control-Allow-Headers', 'Content-Type',true);
             return $response;
-            
         }
-        $newForm = $form->getData();
-        $newNameOfClass=$newForm->getNameClass();
 
-        if($routeEdit==true)
+
+        if($editMode==true)
         {
+            //Permet d'obtenir tous les élèves liés à $nameClass (old) 
+            //et remplacer leur champ 'name_class' par $newNameOfClass (new)
+            $newNameOfClass=$form->getData()->getNameClass();
+
             $entityManager=$this->getDoctrine()->getManager();
             $students=$entityManager->getRepository(Students::class)->findBy(['name_class' => $nameClass]);
             foreach ($students as $student) {
@@ -112,7 +120,6 @@ class ApiController extends AbstractController
                 $entityManager->flush();
             }
         }
-
         $doctrine=$this->getDoctrine()->getManager();
         $doctrine->persist($class);
         $doctrine->flush();
@@ -121,8 +128,7 @@ class ApiController extends AbstractController
             [
                 'status' => 'Class added',
                 'HTTP'=>JsonResponse::HTTP_CREATED
-            ]
-            
+            ]  
         );
         $response->headers->set('Content-Type', 'application/json');
         $response->headers->set('Access-Control-Allow-Origin', '*');
@@ -133,21 +139,37 @@ class ApiController extends AbstractController
     }
     /**
      * @Route("/api/classes/{id}",name="api_deleteClass", methods={"DELETE", "OPTIONS"})
+     * @param int $id id de la classe
      */
     public function APIdeleteClass($id)
     {
+        if ($_SERVER['REQUEST_METHOD'] == 'OPTIONS')
+        {
+            $response= new Response();
+            $response->headers->set('Content-Type', 'application/text');
+            $response->headers->set('Access-Control-Allow-Origin', '*');
+            $response->headers->set("Access-Control-Allow-Methods", "GET, PUT, POST, DELETE, OPTIONS");
+            $response->headers->set('Access-Control-Allow-Headers', 'Content-Type', true);
+            return $response;   
+        }
+        $name_class = $this->getDoctrine()->getManager();
+        $name_class=$name_class->getRepository('App:Classes')->find($id)->getNameClass();
+
+        //Supprime d'abord tous les élèves liés à cette classe
         $entityManager=$this->getDoctrine()->getManager();
-        $deletes=$entityManager->getRepository(Students::class)->findBy(['name_class' => $id]);
+        $deletes=$entityManager->getRepository(Students::class)->findBy(['name_class' => $name_class]);
         foreach ($deletes as $delete) {
             $entityManager->remove($delete);
             $entityManager->flush();
         }
 
-        $deletes=$entityManager->getRepository(Classes::class)->findBy(['name_class' => $id]);
+        $deletes=$entityManager->getRepository(Classes::class)->findBy(['name_class' => $name_class]);
+
         foreach ($deletes as $delete) {
             $entityManager->remove($delete);
             $entityManager->flush();
         }
+    
         $response= new JsonResponse(
             [
                'message' => 'Class deleted', 
@@ -162,6 +184,7 @@ class ApiController extends AbstractController
 
     /**
      * @Route("api/students/{id}",name="api_students", methods={"GET", "OPTIONS"})
+     * @param int $id id de la classe dont on veut observer les élèves
      */
     public function APIstudents($id)
     {
@@ -178,18 +201,14 @@ class ApiController extends AbstractController
         $normalizers = array (new ObjectNormalizer());
         $serializer =  new Serializer($normalizers, $encoders);
 
+        //D'abord le nom de la classe lié à l'id va être recherché
         $name_class = $this->getDoctrine()->getManager();
-        $name_class=$name_class->getRepository('App:Classes')->find($id);
+        $name_class=$name_class->getRepository('App:Classes')->find($id)->getNameClass();
 
+        //Ensuite, on enregistre tous les étudiants dans la liste students
         $em = $this->getDoctrine()->getManager();
         $students=$em->getRepository('App:Students')->findBy(['name_class'=>$name_class]);
-        $var=$name_class->getNameClass();
 
-        if ($var)
-        {
-            $em = $this->getDoctrine()->getManager();
-            $students=$em->getRepository('App:Students')->findBy(['name_class'=>$var]);
-        }
 
         $jsonContent = $serializer->serialize($students, 'json');
         $response = new JsonResponse();
@@ -203,7 +222,8 @@ class ApiController extends AbstractController
 
     /**
      * @Route("api/students",name="api_add_students", methods={"PUT", "OPTIONS"})
-     * *@Route("api/students/{id}/edit", name="api_editStudent", methods={"POST", "OPTIONS"})
+     * @Route("api/students/{id}/edit", name="api_edit_student", methods={"POST", "OPTIONS"})
+     * @param int $id , id de l'étudiant, mode Edit
      */
     public function APIaddStudent(Students $student = null, Request $request)
     {
@@ -216,7 +236,9 @@ class ApiController extends AbstractController
             $response->headers->set('Access-Control-Allow-Headers', 'Content-Type', true);
             return $response;   
         }
+
         $data = json_decode($request->getContent(),true);
+        //Si un étudiant n'a pas été reçu, il faut en créer un nouveau (editMode=false)
         if(!$student)
         {
             $student = new Students();
@@ -252,6 +274,7 @@ class ApiController extends AbstractController
     }
         /**
      * @Route("api/students/{id}",name="api_delete_students", methods={"DELETE", "OPTIONS"})
+     * @param int $id , id de l'étudiant
      */
     public function APIdeleteStudent($id)
     {
@@ -274,9 +297,4 @@ class ApiController extends AbstractController
         $response->headers->set('Access-Control-Allow-Headers', 'Content-Type', true);
         return $response;
     }
-    /* Exemple de JSON
-    {
-	"teacher": "Mr krkr",
-	"name_class" : "Trjkekj"
-        }*/
 }
